requirements.txt
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
PyJWT==1.7.1
pytz==2024.1
PyYAML==6.0.2
setuptools==74.0.0
sqlparse==0.5.1
uritemplate==4.1.1
